microservicio-gemini/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.19
google-auth==2.36.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
requests==2.32.3
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
